Nave bar updated notifications gone and replaced with search removed some boiler code from file too
