+++ app/src/main/res/menu/menu_nav.xml
@@ -9,9 +9,9 @@
         android:icon="@drawable/ic_home"
         android:title="Home" />
     <item
-        android:id="@+id/notifications"
+        android:id="@+id/search"
         android:icon="@drawable/ic_post"
-        android:title="Notifications" />
+        android:title="Search" />
     <item
         android:id="@+id/nav_profile"
         android:icon="@drawable/ic_account"
